modules/direct-link/src/test/java/com/ibm/cloud/networking/direct_link/v1/model/AsPrependTemplateTest.java
/*
 * (C) Copyright IBM Corp. 2022.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */

package com.ibm.cloud.networking.direct_link.v1.model;

import com.ibm.cloud.networking.direct_link.v1.model.AsPrependTemplate;
import com.ibm.cloud.networking.direct_link.v1.utils.TestUtilities;
import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
import java.io.InputStream;
import java.util.HashMap;
import java.util.List;
import org.testng.annotations.Test;
import static org.testng.Assert.*;

/**
 * Unit test class for the AsPrependTemplate model.
 */
public class AsPrependTemplateTest {
  final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
  final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();

  @Test
  public void testAsPrependTemplate() throws Throwable {
    AsPrependTemplate asPrependTemplateModel = new AsPrependTemplate.Builder()
      .length(Long.valueOf("4"))
      .policy("import")
      .prefix("172.17.0.0/16")
      .build();
    assertEquals(asPrependTemplateModel.length(), Long.valueOf("4"));
    assertEquals(asPrependTemplateModel.policy(), "import");
    assertEquals(asPrependTemplateModel.prefix(), "172.17.0.0/16");

    String json = TestUtilities.serialize(asPrependTemplateModel);

    AsPrependTemplate asPrependTemplateModelNew = TestUtilities.deserialize(json, AsPrependTemplate.class);
    assertTrue(asPrependTemplateModelNew instanceof AsPrependTemplate);
    assertEquals(asPrependTemplateModelNew.length(), Long.valueOf("4"));
    assertEquals(asPrependTemplateModelNew.policy(), "import");
    assertEquals(asPrependTemplateModelNew.prefix(), "172.17.0.0/16");
  }

  @Test(expectedExceptions = IllegalArgumentException.class)
  public void testAsPrependTemplateError() throws Throwable {
    new AsPrependTemplate.Builder().build();
  }

}